Avoid overflows by casting to long first (unlikely, but makes warning go away)

+++ plugin/src/main/java/com/playmonumenta/redissync/AutoSaveListener.java
@@ -67,7 +67,7 @@ public class AutoSaveListener implements Listener {
 					}
 
 					mPendingSaves.remove(uuid);
-				}, i * delayBetweenSaves);
+				}, (long) i * delayBetweenSaves);
 
 				mPendingSaves.put(uuid, task);
 			}
